Added 'concise' option and functionality
Specifying -m option or clicking m on a keyboard during execution
gets rid of all unused space and makes the clock be as close to top left corner as possible.
Useful if you want to have the clock in the smallest terminal space possible.

Works (doesn't break) with border/box functionality (Enabling box negates this feature).